app/gateway/requirements.txt
anyio==3.6.2
autoflake==2.0.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
isort==5.12.0
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
passlib==1.7.4
pathspec==0.11.0
pika==1.3.1
platformdirs==3.1.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pydantic==1.10.5
pyflakes==3.0.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
SQLAlchemy-Utils==0.40.0
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
zipp==3.15.0
